+++ modules/AWSPowerShell/AWSPowerShellCompleters.psm1
@@ -28949,6 +28949,17 @@ $MSK_Completers = {
             break
         }
 
+        # Amazon.Kafka.StorageMode
+        {
+            ($_ -eq "New-MSKClusterV2/Provisioned_StorageMode") -Or
+            ($_ -eq "New-MSKCluster/StorageMode") -Or
+            ($_ -eq "Update-MSKStorage/StorageMode")
+        }
+        {
+            $v = "LOCAL","TIERED"
+            break
+        }
+
 
     }
 
@@ -28962,6 +28973,8 @@ $MSK_map = @{
     "EnhancedMonitoring"=@("New-MSKCluster","Update-MSKMonitoring")
     "Provisioned_EncryptionInfo_EncryptionInTransit_ClientBroker"=@("New-MSKClusterV2")
     "Provisioned_EnhancedMonitoring"=@("New-MSKClusterV2")
+    "Provisioned_StorageMode"=@("New-MSKClusterV2")
+    "StorageMode"=@("New-MSKCluster","Update-MSKStorage")
 }
 
 _awsArgumentCompleterRegistration $MSK_Completers $MSK_map
@@ -29048,7 +29061,8 @@ $MSK_SelectMap = @{
                "Update-MSKConfiguration",
                "Update-MSKConnectivity",
                "Update-MSKMonitoring",
-               "Update-MSKSecurity")
+               "Update-MSKSecurity",
+               "Update-MSKStorage")
 }
 
 _awsArgumentCompleterRegistration $MSK_SelectCompleters $MSK_SelectMap
@@ -45461,7 +45475,7 @@ $SM_Completers = {
         # Amazon.SageMaker.HyperParameterTuningJobStrategyType
         "New-SMHyperParameterTuningJob/HyperParameterTuningJobConfig_Strategy"
         {
-            $v = "Bayesian","Hyperband","Random"
+            $v = "Bayesian","Grid","Hyperband","Random"
             break
         }