+++ modules/AWSPowerShell/Cmdlets/IdentityManagement/Basic/New-IAMRole-Cmdlet.cs
@@ -84,7 +84,7 @@ namespace Amazon.PowerShell.Cmdlets.IAM
         /// <para>
         /// <para>The maximum session duration (in seconds) that you want to set for the specified role.
         /// If you do not specify a value for this setting, the default value of one hour is applied.
-        /// This setting can have a value from 1 hour to 12 hours.</para><para>Anyone who assumes the role from the or API can use the <code>DurationSeconds</code>
+        /// This setting can have a value from 1 hour to 12 hours.</para><para>Anyone who assumes the role from the CLI or API can use the <code>DurationSeconds</code>
         /// API parameter or the <code>duration-seconds</code> CLI parameter to request a longer
         /// session. The <code>MaxSessionDuration</code> setting determines the maximum duration
         /// that can be requested using the <code>DurationSeconds</code> parameter. If users don't