+++ modules/AWSPowerShell/Cmdlets/Neptune/Basic/New-NPTDBCluster-Cmdlet.cs
@@ -253,6 +253,30 @@ namespace Amazon.PowerShell.Cmdlets.NPT
         public System.String MasterUserPassword { get; set; }
         #endregion
         
+        #region Parameter ServerlessV2ScalingConfiguration_MaxCapacity
+        /// <summary>
+        /// <para>
+        /// <para>The maximum number of Neptune capacity units (NCUs) for a DB instance in a Neptune
+        /// Serverless cluster. You can specify NCU values in half-step increments, such as 40,
+        /// 40.5, 41, and so on.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public System.Double? ServerlessV2ScalingConfiguration_MaxCapacity { get; set; }
+        #endregion
+        
+        #region Parameter ServerlessV2ScalingConfiguration_MinCapacity
+        /// <summary>
+        /// <para>
+        /// <para>The minimum number of Neptune capacity units (NCUs) for a DB instance in a Neptune
+        /// Serverless cluster. You can specify NCU values in half-step increments, such as 8,
+        /// 8.5, 9, and so on.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public System.Double? ServerlessV2ScalingConfiguration_MinCapacity { get; set; }
+        #endregion
+        
         #region Parameter OptionGroupName
         /// <summary>
         /// <para>
@@ -468,6 +492,8 @@ namespace Amazon.PowerShell.Cmdlets.NPT
             context.PreferredMaintenanceWindow = this.PreferredMaintenanceWindow;
             context.PreSignedUrl = this.PreSignedUrl;
             context.ReplicationSourceIdentifier = this.ReplicationSourceIdentifier;
+            context.ServerlessV2ScalingConfiguration_MaxCapacity = this.ServerlessV2ScalingConfiguration_MaxCapacity;
+            context.ServerlessV2ScalingConfiguration_MinCapacity = this.ServerlessV2ScalingConfiguration_MinCapacity;
             context.StorageEncrypted = this.StorageEncrypted;
             if (this.Tag != null)
             {
@@ -589,6 +615,35 @@ namespace Amazon.PowerShell.Cmdlets.NPT
             {
                 request.ReplicationSourceIdentifier = cmdletContext.ReplicationSourceIdentifier;
             }
+            
+             // populate ServerlessV2ScalingConfiguration
+            var requestServerlessV2ScalingConfigurationIsNull = true;
+            request.ServerlessV2ScalingConfiguration = new Amazon.Neptune.Model.ServerlessV2ScalingConfiguration();
+            System.Double? requestServerlessV2ScalingConfiguration_serverlessV2ScalingConfiguration_MaxCapacity = null;
+            if (cmdletContext.ServerlessV2ScalingConfiguration_MaxCapacity != null)
+            {
+                requestServerlessV2ScalingConfiguration_serverlessV2ScalingConfiguration_MaxCapacity = cmdletContext.ServerlessV2ScalingConfiguration_MaxCapacity.Value;
+            }
+            if (requestServerlessV2ScalingConfiguration_serverlessV2ScalingConfiguration_MaxCapacity != null)
+            {
+                request.ServerlessV2ScalingConfiguration.MaxCapacity = requestServerlessV2ScalingConfiguration_serverlessV2ScalingConfiguration_MaxCapacity.Value;
+                requestServerlessV2ScalingConfigurationIsNull = false;
+            }
+            System.Double? requestServerlessV2ScalingConfiguration_serverlessV2ScalingConfiguration_MinCapacity = null;
+            if (cmdletContext.ServerlessV2ScalingConfiguration_MinCapacity != null)
+            {
+                requestServerlessV2ScalingConfiguration_serverlessV2ScalingConfiguration_MinCapacity = cmdletContext.ServerlessV2ScalingConfiguration_MinCapacity.Value;
+            }
+            if (requestServerlessV2ScalingConfiguration_serverlessV2ScalingConfiguration_MinCapacity != null)
+            {
+                request.ServerlessV2ScalingConfiguration.MinCapacity = requestServerlessV2ScalingConfiguration_serverlessV2ScalingConfiguration_MinCapacity.Value;
+                requestServerlessV2ScalingConfigurationIsNull = false;
+            }
+             // determine if request.ServerlessV2ScalingConfiguration should be set to null
+            if (requestServerlessV2ScalingConfigurationIsNull)
+            {
+                request.ServerlessV2ScalingConfiguration = null;
+            }
             if (cmdletContext.StorageEncrypted != null)
             {
                 request.StorageEncrypted = cmdletContext.StorageEncrypted.Value;
@@ -686,6 +741,8 @@ namespace Amazon.PowerShell.Cmdlets.NPT
             public System.String PreferredMaintenanceWindow { get; set; }
             public System.String PreSignedUrl { get; set; }
             public System.String ReplicationSourceIdentifier { get; set; }
+            public System.Double? ServerlessV2ScalingConfiguration_MaxCapacity { get; set; }
+            public System.Double? ServerlessV2ScalingConfiguration_MinCapacity { get; set; }
             public System.Boolean? StorageEncrypted { get; set; }
             public List<Amazon.Neptune.Model.Tag> Tag { get; set; }
             public List<System.String> VpcSecurityGroupId { get; set; }